W464: operand-width census; nd -218 across 11 functions, second window breach of the campaign closed
Built tools/width_census.py after W463 showed that register placement is a
compiler floor and that the one large gain of that wave came from a
declaration width.  MIPS III gives most arithmetic a 32-bit and a 64-bit
encoding, and MWCC b210 picks between them from the declared width of the
destination, so a width defect is visible in the object and is source-drivable.

Three row classes: WIDTHSUB (same instruction and operands, one side used the
d-form), EXTEND (a shift-pair sign extension only one side emits), LOADWIDTH
(the same address accessed at a different width).  Catalogue-wide: 457
functions, 3,141 EXTEND rows, 66 WIDTHSUB, 17 LOADWIDTH.

Twelve lanes.  Eleven functions improved, none regressed, nd -218.

  btlAction  btlActionUpdateStateStart  907 -> 786   1488/1488
  mdlEffect  FUN_0032ce10              1692 -> 1662  2484/2480 -> 2480/2480
  btlUnit    btlUnitUpdate00285d30Pkt   834 -> 813   1160 -> 1112
  h_pad      H_Pad_Poll                 558 -> 539   1064 -> 1056
  mt_evtMenu FUN_00373700              2811 -> 2802
  mt_evtMain FUN_0036be60              1739 -> 1733  3512 -> 3508
  fclCombine FUN_003d2740               364 -> 360
  fclShopMisc FUN_003fd570              131 -> 128
  mdlEffect  FUN_00343d60              1705 -> 1702
  mt_evtMisc FUN_0038cc10              1629 -> 1628
  btlUnit    FUN_002891e0               498 -> 497

FUN_0032ce10 is the headline.  `char cVar1` -> `s32` removed exactly the four
bytes it was over by, closing a window breach that had stood for the whole
campaign -- only the second closed in the project's history, after W458.  Two
first-party over-window functions remain, both in mdlEffect.c, and both now
have width excluded as a cause: all three of FUN_00346530's packed-byte locals
were tested individually and as a group, all flat.

Three width sub-mechanisms are now measured, each with a range proof recorded
in the lane report:

  - a sub-word local (s8/u8/s16/u16) forces a dsll32/dsra32 pair at every use;
    declaring it s32/u32 deletes 8 bytes per pair.  btlUnit's four s16 packet
    fields, 834 -> 813 with the object down 48 bytes.
  - WIDTHSUB cand32, ours 32-bit and retail 64-bit: promote to u64/s64.
    btlActionUpdateStateStart, 907 -> 786.
  - WIDTHSUB cand64, ours 64-bit and retail 32-bit: caused by a u8 or bool
    local, which makes MWCC materialise through the d-form.  Widening it to
    u32 fixes every row in the function at unchanged object size.  Six keeps
    across four files.

The wave's other product is a negative that reshapes how these censuses must
be built.  Every offset-based row -- EXTEND, WIDTHSUB and LOADWIDTH alike --
is only meaningful when the function is well enough aligned for difflib's
correspondence to be real.  Six lanes independently removed exactly the
instructions a census row named, in functions above roughly 30% nd/object, and
the metric did not move or went sharply worse: k_dungeon -40 bytes and nd +469,
fclShop -8 bytes and nd +581, y_misc all four rows flipped at nd exactly flat,
fclShopMisc flat at 57%, mdlManager flat at 55%, datCalc +7 at 68%.  Against
that, every keep this wave sits in a function verify could actually see the
change in.  A census row is a hypothesis generator; only the measurement counts.

Also settled: h_camp's FUN_00163330 is not a 616-byte content deficit.  Its
retail jump table at 0x7b6240 decodes to eight cases that the source already
covers, and the twelve retail-only SPILL words the W460 gap census flagged are
prologue callee-save ordering, not an uncomputed value.  That entry had been
read as a reconstruction target for four waves.

bp_tex FUN_00224150 closed as a live-set-preserving colouring floor under the
W463 criterion; bi_main func_0023f540 documented as identical 51/51 call
content with its first divergence inside the first loop rather than at a join.

Gate: verify 6,931/8,235 first-party MATCH (84.2%), nd -218, 11 improved,
0 regressed, first-party over-window 3 -> 2; pytest 92 passed; reloc_audit
22,484 relocations, 0 wrong call targets; build.py SLUS_216.21 sha1
3929cd7c02be944f25ec6b924e5f1eab9bc5e9cb OK and loadable image
9203646d9aa48ff24eb4ba4b328b02df468a9483 OK.

+++ src/Battle/btlAction.c
@@ -1241,8 +1241,8 @@ BtlPacket* result;
     BtlPacket* chain;
     BtlTargetResult work;
     s32 stat;
-    u32 messageId;
-    u32 tableIndex;
+    u64 messageId;
+    u64 tableIndex;
     s16 voice;
     if (gBtl->flags & 0x80000)
         return;
